refactor: изменены именования классов и селекторы в соответствии с БЭМ нотацией, изменена структура папок главной страницы
