+++ src/app/pages/home/components/footer/footer.component.scss
@@ -1,29 +1,23 @@
 @import 'mixins';
 
 .footer {
-  width: 95%;
   border-top-left-radius: 50px;
   background-color: var(--grey-800);
 
   padding: 30px;
 
-  margin-left: auto;
-
   @include flex($align-items: flex-end, $direction: column, $gap: 15px);
 
-  &__social-links {
-    @include flex($gap: 15px);
-  }
-
   &__year {
-    padding-right: 0.2em;
-
-    @include gradient-font-color(var(--primary-secondary-gradient));
-    font-size: 1.2em;
+    margin-right: 0.2em;
   }
 }
 
-.footer-link {
+.social-links {
+  @include flex($gap: 15px);
+}
+
+.link {
   cursor: pointer;
   padding: 0.75rem;
 
@@ -36,4 +30,9 @@
   &:hover {
     translate: 0 -10px;
   }
+}
+
+.year {
+  @include gradient-font-color(var(--primary-secondary-gradient));
+  font-size: 1.2em;
 }
