+++ src/app/pages/home/components/header/header.component.html
@@ -8,9 +8,9 @@
       <img
         src="../../../../../assets/images/logo.svg"
         alt="Логотип"
-        class="header__logo"
+        class="header__logo logo"
       />
-      <h1 class="header__heading">Финансы</h1>
+      <h1 class="header__heading heading">Финансы</h1>
     </ng-template>
   </p-menubar>
 </header>
