+++ src/app/pages/home/components/home-entry/home-entry.component.html
@@ -1,5 +1,5 @@
-<app-header></app-header>
+<app-header class="header"></app-header>
 <main class="main">
   <router-outlet></router-outlet>
 </main>
-<app-footer></app-footer>
+<app-footer class="footer"></app-footer>
