+++ src/app/pages/home/components/home-page/home-currencies-section/home-currencies-section.component.html
@@ -1,8 +1,8 @@
 <section class="currencies-section">
-  <div class="currencies-section__heading-wrapper">
-    <span class="currencies-section__heading-emoji">📈</span>
-    <h2 class="currencies-section__heading">Изучите Пульс Криптовалюты</h2>
-    <span class="currencies-section__heading-emoji">📉</span>
+  <div class="currencies-section__heading heading">
+    <span class="heading__emoji">📈</span>
+    <h2 class="heading__main-text">Изучите Пульс Криптовалюты</h2>
+    <span class="heading__emoji">📉</span>
   </div>
   <p class="currencies-section__description">
     В этой секции вы найдете графики курсов криптовалют, которые позволят вам
