+++ src/app/pages/home/components/home-page/home-diversification-section/home-diversification-section.component.html
@@ -1,10 +1,8 @@
 <section class="diversification-section">
-  <div class="diversification-section__text-block">
-    <h2 class="diversification-section__heading">
-      Мир банковских услуг изменился
-    </h2>
-    <p class="diversification-section__main-text">
-      <span class="diversification-section__emphasized-text"
+  <div class="diversification-section__text-block text-block">
+    <h2 class="text-block__heading heading">Мир банковских услуг изменился</h2>
+    <p class="text-block__main-text">
+      <span class="text-block__emphasized-text emphasized-text"
         >Диверсификация является стратегическим приоритетом.</span
       >
       Финансовым учреждениям необходимо повышать уровень потребительских и малых
