+++ src/app/pages/home/components/home-page/home-page.component.scss
