+++ src/app/pages/home/home.routes.ts
@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import { HomePageComponent } from './components/home-page/home-page/home-page.component';
+import { HomePageComponent } from './components/home-page/home-page.component';
 import { HomeEntryComponent } from './components/home-entry/home-entry.component';
 
 const routes: Routes = [
